consolidate test files @sanderegg
